Tweak sample app and fix issue with now line being drawn at the wrong time

+++ library/src/main/java/com/alamkanak/weekview/WeekViewConfigWrapper.kt
@@ -426,7 +426,7 @@ internal class WeekViewConfigWrapper(
         get() = config.maxHour - config.minHour
 
     val minutesPerDay: Int
-        get() = hoursPerDay * Constants.MINUTES_PER_HOUR
+        get() = (hoursPerDay * Constants.MINUTES_PER_HOUR).toInt()
 
     val startHour: Int
         get() = if (showMidnightHour && showTimeColumnHourSeparator) minHour else timeColumnHoursInterval
